refactor(ppvm-lindblad): split lib.rs along spec/basis/step lines (#215)
## Summary
- Same kind of cleanup as #193 did for #180: split the 1300-line
`ppvm-lindblad` `lib.rs` / `LindbladSpec` impl into `word`, `algebra`,
`spec`, `basis`, `step`, and `tests`, keeping the public
`ppvm_lindblad::*` API.
- Collapses the timed/untimed `pc_step` copy, demotes `ppvm-pauli-sum`
to a dev-dependency, drops unused `approx`, strips 182-forward
`orbit_rep` doc links, and forwards `admit_basis` / `tau_add` on the
Python string `pc_step`.
- **Base is #181 (`split/2-ctpp-core`) on purpose.** #193 replaces #180
and is not an ancestor of #181; retargeting this at #193 now would mix
the whole CTPP crate into the diff. After #193 merges and #181 is
rebased onto it, this PR rebases with #181.

## Test plan
- [x] `cargo test -p ppvm-lindblad` (6 passed)
- [x] `cargo clippy -p ppvm-lindblad --all-targets -- -D warnings`
- [x] `cargo check -p ppvm-python-native`
- [ ] Rebase onto #181 after #181 is retargeted onto #193


Made with [Cursor](https://cursor.com)

---------
