+++ crates/ppvm-lindblad/Cargo.toml
@@ -10,7 +10,6 @@ ndarray = "0.17"
 num = "0.4.3"
 ppvm-traits     = { version = "0.1.0", path = "../ppvm-traits" }
 ppvm-pauli-word = { version = "0.1.0", path = "../ppvm-pauli-word" }
-ppvm-pauli-sum  = { version = "0.1.0", path = "../ppvm-pauli-sum" }
 rayon = "1.11"
 # Matrix-exponential action (Al-Mohy & Higham). QuSpin-rust is MIT-licensed;
 # the pinned rev is the commit that added the LICENSE file.
@@ -21,4 +20,4 @@ quspin-expm = { git = "https://github.com/QuSpin/QuSpin-rust", rev = "a0ad6c9fe2
 quspin-types = { git = "https://github.com/QuSpin/QuSpin-rust", rev = "a0ad6c9fe2e8063208f9ba1c6677150c993bb554" }
 
 [dev-dependencies]
-approx = "0.5.1"
+ppvm-pauli-sum = { version = "0.1.0", path = "../ppvm-pauli-sum" }
